abstract out stream related utils
