+++ packages/next/server/web/utils.ts
@@ -14,32 +14,6 @@ export async function* streamToIterator<T>(
   reader.releaseLock()
 }
 
-export function readableStreamTee<T = any>(
-  readable: ReadableStream<T>
-): [ReadableStream<T>, ReadableStream<T>] {
-  const transformStream = new TransformStream()
-  const transformStream2 = new TransformStream()
-  const writer = transformStream.writable.getWriter()
-  const writer2 = transformStream2.writable.getWriter()
-
-  const reader = readable.getReader()
-  function read() {
-    reader.read().then(({ done, value }) => {
-      if (done) {
-        writer.close()
-        writer2.close()
-        return
-      }
-      writer.write(value)
-      writer2.write(value)
-      read()
-    })
-  }
-  read()
-
-  return [transformStream.readable, transformStream2.readable]
-}
-
 export function notImplemented(name: string, method: string): any {
   throw new Error(
     `Failed to get the '${method}' property on '${name}': the property is not implemented`
